fix a bug in CachedOp. (#11675)
* fix a bug.

* add tests.

* use default context.

* move all tests to test_contrib_control_flow.py

* fix test.
